#22 code refactored in tests

+++ clients/user-services/spring-data-jpa-user-service/src/main/java/com/github/ymikevich/spring/data/jpa/user/service/service/SpringDataJpaUserService.java
@@ -55,7 +55,7 @@ public class SpringDataJpaUserService implements UserService {
 
     @Override
     public void deleteAccountById(Long id) {
-        accountRepository.deleteById(id);
+        accountRepository.deleteById(id.intValue());
     }
 
     @Override
